+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 ### Fixed
 - Fixed deprecation errors in templates. ([#25](https://github.com/craftcms/digital-products/issues/25))
 - Fixed a bug where deleting a product was impossible from its page. ([#24](https://github.com/craftcms/digital-products/issues/24))
+- Fixed an incorrect plugin settings URL. ([#23](https://github.com/craftcms/digital-products/issues/23))
 
 ## 2.0.3 - 2018-10-19